+++ app/Http/Controllers/Geo/Currencies/Available.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace App\Http\Controllers\Geo\Countries;
+namespace App\Http\Controllers\Geo\Currencies;
 
 use App\Http\Controllers\Controller;
-use Enraiged\Geo\Models\Country;
+use Enraiged\Geo\Models\Currency;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 
@@ -15,16 +15,16 @@ class Available extends Controller
      */
     public function __invoke(Request $request): JsonResponse
     {
-        $countries = Country::query()
+        $currencies = Currency::query()
             ->where('is_active', true)
             ->get()
-            ->transform(fn ($country) => [
-                'id' => $country->id,
-                'code' => $country->code,
-                'name' => $country->name,
+            ->transform(fn ($currency) => [
+                'id' => $currency->id,
+                'code' => $currency->code,
+                'name' => $currency->name,
             ])
             ->toArray();
 
-        return response()->json($countries);
+        return response()->json($currencies);
     }
 }