+++ database/migrations/2023_12_13_143000_create_currencies_table.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     *  Run the migrations.
+     *
+     *  @return void
+     */
+    public function up()
+    {
+        Schema::create('currencies', function (Blueprint $table) {
+            $table->id();
+            $table->char('code', 3)->unique();
+            $table->string('name', 32);
+            $table->string('symbol', 1);
+            $table->boolean('is_active')->default(true);
+        });
+    }
+
+    /**
+     *  Reverse the migrations.
+     *
+     *  @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('currencies');
+    }
+};