+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 2026-04-22 : 0.4.0-rc.2  
 Update: Testing,preparation in enraiged-laravel for initial release
 
+- Added currencies migration,seeding to the database
 - Added email column to the profiles table migration
 - Added is_{billable,deliverable} columns to the addresses
 - Corrected issue with missing country_id when creating user